service/requirements.txt
AWSIoTPythonSDK==1.4.8
dbus-python==1.2.16
Jinja2==2.11.2
pbkdf2==1.3
pycountry==19.8.18
